+++ .gitignore
@@ -1,24 +1,28 @@
-# build output
-dist/
-# generated types
-.astro/
+dist
+.solid
+.output
+.vercel
+.netlify
+.vinxi
+app.config.timestamp_*.js
+
+# Environment
+.env
+.env*.local
 
 # dependencies
-node_modules/
+/node_modules
 
-# logs
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-pnpm-debug.log*
+# IDEs and editors
+/.idea
+.project
+.classpath
+*.launch
+.settings/
 
+# Temp
+gitignore
 
-# environment variables
-.env
-.env.production
-
-# macOS-specific files
+# System Files
 .DS_Store
-
-# jetbrains setting folder
-.idea/
+Thumbs.db
