--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-	"editor.defaultFormatter": "biomejs.biome"
-}
