+++ package.json
@@ -1,37 +1,29 @@
 {
-	"name": "@specta-rs/website",
-	"type": "module",
-	"version": "0.0.0",
-	"private": true,
-	"scripts": {
-		"dev": "astro dev",
-		"start": "astro dev",
-		"build": "astro check && astro build",
-		"preview": "astro preview",
-		"astro": "astro",
-		"format": "pnpm biome check --write ."
-	},
-	"dependencies": {
-		"@astrojs/check": "^0.9.4",
-		"@astrojs/solid-js": "^4.4.3",
-		"@astrojs/tailwind": "^5.1.2",
-		"@fontsource-variable/raleway": "^5.1.0",
-		"@fontsource/raleway": "^5.1.0",
-		"@solidjs/router": "^0.13.6",
-		"astro": "^4.16.13",
-		"clsx": "^2.1.1",
-		"satori": "^0.10.14",
-		"solid-js": "^1.9.3",
-		"tailwindcss": "^3.4.15",
-		"typescript": "^5.6.3"
-	},
-	"devDependencies": {
-		"@astrojs/sitemap": "^3.2.1",
-		"@biomejs/biome": "^1.9.4",
-		"@tailwindcss/typography": "^0.5.15",
-		"@types/node": "^20.17.6",
-		"sharp": "^0.33.5",
-		"vite": "^5.4.11",
-		"vite-plugin-arraybuffer": "^0.0.8"
-	}
+  "name": "@specta-rs/website",
+  "type": "module",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "dev": "vinxi dev",
+    "build": "vinxi build",
+    "start": "vinxi start",
+    "version": "vinxi version",
+    "format": "pnpm biome check --write ."
+  },
+  "dependencies": {
+    "@kobalte/core": "^0.13.7",
+    "@kobalte/solidbase": "npm:@brendonovich/kobalte__solidbase@^0.0.20",
+    "@solid-primitives/event-listener": "^2.3.3",
+    "@solid-primitives/keyboard": "^1.2.8",
+    "@solid-primitives/platform": "^0.1.2",
+    "@solidjs/meta": "^0.29.4",
+    "@solidjs/router": "^0.15.1",
+    "@solidjs/start": "^1.0.10",
+    "solid-devtools": "^0.30.1",
+    "solid-js": "^1.9.3",
+    "vinxi": "^0.4.3"
+  },
+  "engines": {
+    "node": ">=18"
+  }
 }
