+++ public/images/tauri-specta-logo.png
