--- solid-project/package.json
@@ -1,26 +1,0 @@
-{
-	"name": "example-basic",
-	"type": "module",
-	"scripts": {
-		"dev": "vinxi dev",
-		"build": "vinxi build",
-		"start": "vinxi start",
-		"version": "vinxi version"
-	},
-	"dependencies": {
-		"@kobalte/core": "^0.13.7",
-		"@kobalte/solidbase": "npm:@brendonovich/kobalte__solidbase@^0.0.20",
-		"@solid-primitives/event-listener": "^2.3.3",
-		"@solid-primitives/keyboard": "^1.2.8",
-		"@solid-primitives/platform": "^0.1.2",
-		"@solidjs/meta": "^0.29.4",
-		"@solidjs/router": "^0.15.1",
-		"@solidjs/start": "^1.0.10",
-		"solid-devtools": "^0.30.1",
-		"solid-js": "^1.9.3",
-		"vinxi": "^0.4.3"
-	},
-	"engines": {
-		"node": ">=18"
-	}
-}
