--- solid-project/src/app.tsx
@@ -1,11 +1,0 @@
-import { Router } from "@solidjs/router";
-import { FileRoutes } from "@solidjs/start/router";
-import { SolidBaseRoot } from "@kobalte/solidbase/client";
-
-export default function App() {
-	return (
-		<Router root={SolidBaseRoot}>
-			<FileRoutes />
-		</Router>
-	);
-}
