--- solid-project/src/routes/index.mdx
@@ -1,15 +1,0 @@
----
-title: specta-rs
-layout: home
-hero:
-  text: Rust crates for building better web apps
-  image:
-    src: https://specta.dev/_astro/logo.QnjcQVw0.png
-features:
-  - icon: <img src="https://specta.dev/_astro/logo.QnjcQVw0.png" height="80"/>
-    title: specta
-  - icon: <img src="https://specta.dev/_astro/rspc-logo.5MU6p3sA.png" height="80"/>
-    title: rspc
-  - icon: <img src="https://specta.dev/_astro/tauri-specta-logo.Cnl2of5a.png" height="80"/>
-    title: tauri-specta
----
