--- src/content/docs/project/mission.md
@@ -1,7 +1,0 @@
----
-title: Mission
----
-
-# TODO
-
-Mission Page
