--- src/content/docs/tauri-specta/todo.md
@@ -1,7 +1,0 @@
----
-title: Overview
----
-
-# TODO
-
-Tauri Specta Overview
