--- src/pages/[...slug].astro
@@ -1,98 +1,0 @@
----
-import { App, getRoutes } from "../app";
-import { getCollection } from "astro:content";
-import type { Navigation } from "../util/navctx";
-import "@fontsource-variable/raleway";
-
-const navbarTitle: Record<string, string> = {
-	project: "Project",
-	"tauri-specta": "Tauri Specta",
-};
-
-async function generateNavigation() {
-	const docs = await getCollection("docs");
-	const result: Record<string, Navigation[number]> = {};
-	for (const doc of docs) {
-		const [headingSlug] = doc.slug.split("/");
-		if (!(headingSlug in result))
-			result[headingSlug] = {
-				title: navbarTitle?.[headingSlug] ?? headingSlug,
-				links: [],
-			};
-
-		result[headingSlug].links.push({
-			title: doc.data.title,
-			href: doc.slug,
-		});
-	}
-	return Object.values(result);
-}
-
-export async function getStaticPaths() {
-	return getRoutes().map((slug) => ({
-		params: {
-			slug,
-		},
-	}));
-}
-
-const navigation = await generateNavigation();
-
-// When using `build { format: "file" }` the prerender will target `.html` which is not what Solid Router expects.
-// In production Cloudflare Pages will strip the `.html`
-const path = Astro.url.pathname.endsWith(".html")
-	? Astro.url.pathname.slice(0, -5)
-	: Astro.url.pathname;
----
-
-<html lang="en" class="h-full w-full">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" type="image/x-icon" href="/favicon.ico" />
-    <meta name="viewport" content="width=device-width" />
-    <meta name="generator" content={Astro.generator} />
-    <title>specta-rs</title>
-    <meta property="description" content="Rust crates for building better web apps" />
-    <meta name="author" content="Oscar Beaumont" />
-    <link rel="canonical" href={Astro.site!.origin} />
-    
-    <meta property="og:title" content="specta-rs" />
-    <meta property="og:description" content="Rust crates for building better web apps" />
-    <meta property="og:url" content={Astro.site!.origin} />
-    <meta property="og:image" content={`${Astro.site!.origin}/opengraph.png`} />
-    <meta property="og:image:type" content="image/jpeg" />
-    <meta property="og:image:width" content="1600" />
-    <meta property="og:image:height" content="800" />
-
-    <link rel="sitemap" href="/sitemap-index.xml" />
-
-    <style>
-      body {
-        background-color: #0e111a;
-        background-image: linear-gradient(
-            to bottom,
-            transparent 90%,
-            rgb(30 41 59) 10%
-          ),
-          linear-gradient(to right, #0e111a 90%, rgb(30 41 59) 10%);
-        background-size:
-          5px 5px,
-          5px 5px;
-
-        font-family: "Raleway", sans-serif;
-        font-optical-sizing: auto;
-        font-weight: 600;
-        font-style: normal;
-        color: red;
-      }
-
-      /* Dark mode scroll bars */
-      :root {
-        color-scheme: dark;
-      }
-    </style>
-  </head>
-  <body class="h-full w-full text-white pt-16 overflow-y-hidden">
-    <App path={path} navigation={navigation} client:idle />
-  </body>
-</html>
