--- src/pages/robots.txt.ts
@@ -1,16 +1,0 @@
-import type { APIRoute } from "astro";
-
-const robotsTxt = `
-User-agent: *
-Allow: /
-
-Sitemap: ${new URL("sitemap-index.xml", import.meta.env.SITE).href}
-`.trim();
-
-export const GET: APIRoute = () => {
-	return new Response(robotsTxt, {
-		headers: {
-			"Content-Type": "text/plain; charset=utf-8",
-		},
-	});
-};
