--- src/routes.gen.tsx
@@ -1,9 +1,0 @@
-// @ts-nocheck
-// DO NOT MODIFY THIS FILE. It's generated by a Vite plugin!
-
-export const manifest = {
-	"/tauri-specta/todo": () => import("./content/docs/tauri-specta/todo.md?markdown"),
-	"/tauri-specta/v1": () => import("./content/docs/tauri-specta/v1.md?markdown"),
-	"/tauri-specta/v2": () => import("./content/docs/tauri-specta/v2.md?markdown"),
-	"/project/mission": () => import("./content/docs/project/mission.md?markdown"),
-} satisfies Record<string, () => unknown>;
