+++ src/routes/[...404].tsx
