+++ src/routes/index.mdx
@@ -1,0 +1,15 @@
+---
+title: specta-rs
+layout: home
+hero:
+  text: Rust crates for building better web apps
+  image:
+    src: /public/logo.png
+features:
+  - icon: <img src="/public/logo.png" height="80"/>
+    title: specta
+  - icon: <img src="/public/rspc-logo.png" height="80"/>
+    title: rspc
+  - icon: <img src="/public/tauri-specta-logo.png" height="80"/>
+    title: tauri-specta
+---
