--- tailwind.config.mjs
@@ -1,8 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-	content: ["./src/**/*.{astro,html,js,jsx,md,mdx,ts,tsx}"],
-	theme: {
-		extend: {},
-	},
-	plugins: [require("@tailwindcss/typography")],
-};
